fix: ST_NumInteriorRing failing with wrong return type (#15202)
Summary:
ST_NumInteriorRing return type should be bigint based on : https://github.com/prestodb/presto/issues/26343

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15202

Reviewed By: bikramSingh91

Differential Revision: D85060783

Pulled By: pedroerp

fbshipit-source-id: 3faaf3edbcd046d05f50591092cc39d998f4e3f2

+++ velox/functions/prestosql/tests/GeometryFunctionsTest.cpp
@@ -2010,8 +2010,8 @@ TEST_F(GeometryFunctionsTest, testStInteriorRingN) {
 TEST_F(GeometryFunctionsTest, testStNumInteriorRing) {
   const auto testStNumInteriorRingFunc =
       [&](const std::optional<std::string>& wkt,
-          const std::optional<int32_t>& expected) {
-        std::optional<int32_t> result = evaluateOnce<int32_t>(
+          const std::optional<int64_t>& expected) {
+        std::optional<int64_t> result = evaluateOnce<int64_t>(
             "ST_NumInteriorRing(ST_GeometryFromText(c0))", wkt);
 
         if (expected.has_value()) {
